libs/product/src/services/adapter/default-product.adapter.ts
import { HttpService, JsonAPITransformerService } from '@spryker-oryx/core';
import { inject } from '@spryker-oryx/injector';
import { Observable, switchMap } from 'rxjs';
import { ApiModel, Product, ProductQualifier } from '../../models';
import { GlueProduct, ProductNormalizer } from './normalizers';
import { ProductAdapter } from './product.adapter';

export class DefaultProductAdapter implements ProductAdapter {
  protected productEndpoint = 'concrete-products';

  constructor(
    protected http = inject(HttpService),
    protected SCOS_BASE_URL = inject('SCOS_BASE_URL'),
    protected transformer = inject(JsonAPITransformerService)
  ) {}

  getKey(qualifier: ProductQualifier): string {
    return (qualifier.sku ?? '') + qualifier.include?.sort()?.join('');
  }

  get({ sku, include }: ProductQualifier): Observable<Product> {
    include = [...Object.values(ApiModel.INCLUDES), ...(include ?? [])].filter(
      (type, index, arr) => arr.indexOf(type) === index
    );

    return this.http
      .get<GlueProduct>(
        `${this.SCOS_BASE_URL}/${this.productEndpoint}/${sku}${
          include ? '?include=' : ''
        }${include?.join(',') || ''}`
      )
      .pipe(
        switchMap((res) =>
          this.transformer.transform<Product>(res, ProductNormalizer)
        )
      );
  }
}
